pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["sqlsense*"]

[project]
name = "sqlsense"
version = "0.1.0"
description = "Safe, audited SQL for AI agents via MCP"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
authors = [{ name = "SQLSense Contributors" }]
keywords = ["mcp", "sql", "ai", "agents", "guardrails", "llm", "claude", "database"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = []  # zero mandatory deps — connectors are optional

[project.optional-dependencies]
postgres   = ["psycopg2-binary>=2.9"]
sqlserver  = ["pyodbc>=4.0"]
snowflake  = ["snowflake-connector-python>=3.0"]
all        = ["psycopg2-binary>=2.9", "pyodbc>=4.0", "snowflake-connector-python>=3.0"]
dev        = ["pytest>=7", "pytest-cov", "ruff", "mypy"]

[project.scripts]
sqlsense = "sqlsense.cli:main"

[project.urls]
Homepage   = "https://github.com/yourusername/sqlsense"
Repository = "https://github.com/yourusername/sqlsense"
Issues     = "https://github.com/yourusername/sqlsense/issues"
Docs       = "https://github.com/yourusername/sqlsense#readme"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts   = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
